Move 'b' function from triag to iterative delete
